main: Move waitready to own file
